- database.dart: included both csv_presets table creation and pressure
  column INTEGER-to-REAL migration in the from<58 block
- export_providers.dart: kept our deletion of legacy importDivesFromCsv
  (replaced by universal import wizard)
- csv_import_service.dart: kept our deletion (legacy, replaced)
- Took main's DiveTank pressure type changes (int? -> double?)
- Updated tank_extractor to preserve pressure precision as double
- Regenerated Drift database code
- Took main's repository changes for pressure type compatibility

+++ integration_test/helpers/screenshot_test_data.dart
@@ -447,9 +447,9 @@ class ScreenshotTestDataSeeder {
               id: _uuid.v4(),
               diveId: diveId,
               volume: const Value(11.1),
-              workingPressure: const Value(207),
-              startPressure: Value(200 + random.nextInt(10)),
-              endPressure: Value(40 + random.nextInt(30)),
+              workingPressure: const Value(207.0),
+              startPressure: Value((200 + random.nextInt(10)).toDouble()),
+              endPressure: Value((40 + random.nextInt(30)).toDouble()),
               o2Percent: const Value(21.0),
             ),
           );
